feat: Show error title for okta verify version upgrade Resolves: OKTA-373720
