+++ packages/@okta/i18n/src/properties/login.properties
@@ -841,6 +841,9 @@ oie.on_prem.verify.passcode.label = Enter code
 oie.numberchallenge.warning = Haven't received a push notification yet? Try opening the Okta Verify app on your device, or <$1>resend the push notification</$1>.
 ## {0} refers to a number which is the correct answer for the number challenge
 oie.numberchallenge.instruction = On your mobile device, open the Okta Verify prompt, then tap <$1>{0}</$1> in Okta Verify to continue.
+oie.numberchallenge.force.upgrade.title = Update Okta Verify
+oie.numberchallenge.force.upgrade = Your response was received, but your Okta Verify version is no longer supported by your organization. To verify your identity with push notifcations, update Okta Verify to the latest version, then try again.
+
 
 ## RSA Authenticator
 oie.rsa.label = RSA SecurID
