+++ src/v2/view-builder/views/ov/ChallengeOktaVerifyResendPushView.js
@@ -3,6 +3,8 @@ import { BaseForm } from '../../internals';
 import BaseAuthenticatorView from '../../components/BaseAuthenticatorView';
 import AuthenticatorVerifyFooter from '../../components/AuthenticatorVerifyFooter';
 
+const OV_NMC_FORCE_UPGRAGE_SERVER_KEY = 'idx.authenticator.app.method.push.force.upgrade.number_challenge';
+
 const Body = BaseForm.extend(Object.assign(
   {
     className: 'okta-verify-resend-push',
@@ -25,10 +27,17 @@ const Body = BaseForm.extend(Object.assign(
         messagesObjs.value.forEach(messagesObj => {
           const msg = messagesObj.message;
           if (messagesObj?.class === 'ERROR') {
-            this.add(createCallout({
+            const options = {
               content: msg,
               type: 'error',
-            }), '.o-form-error-container');
+            };
+            if (this.options.appState.containsMessageWithI18nKey(OV_NMC_FORCE_UPGRAGE_SERVER_KEY)) {
+              // account for error customization
+              options.content = loc('oie.numberchallenge.force.upgrade', 'login');
+              // add a title for OV force upgrade
+              options.title = loc('oie.numberchallenge.force.upgrade.title', 'login');
+            }
+            this.add(createCallout(options), '.o-form-error-container');
           } else {
             this.add(`<p>${msg}</p>`, '.ion-messages-container');
           }
