+++ test/testcafe/framework/page-objects/ChallengeOktaVerifyPushPageObject.js
@@ -3,6 +3,7 @@ import ChallengeFactorPageObject from './ChallengeFactorPageObject';
 const FORM_INFOBOX_WARNING = '.okta-form-infobox-warning';
 const FORM_INFOBOX_ERROR = '.o-form-error-container .infobox-error';
 const RESEND_NUMBER_CHALLENGE_BUTTON = '.okta-form-infobox-warning .resend-number-challenge';
+const FORM_INFOBOX_ERROR_TITLE = '.o-form-error-container .infobox-error > h3';
 
 export default class ChallengeOktaVerifyPushPageObject extends ChallengeFactorPageObject {
   constructor(t) {
@@ -35,6 +36,10 @@ export default class ChallengeOktaVerifyPushPageObject extends ChallengeFactorPa
 
   getErrorBox () {
     return this.form.getElement(FORM_INFOBOX_ERROR);
+  }
+
+  getErrorTitle() {
+    return this.form.getElement(FORM_INFOBOX_ERROR_TITLE);
   }
 
   getWarningBox () {
