+++ test/testcafe/spec/ChallengeOktaVerifyPush_spec.js
@@ -5,7 +5,6 @@ import { checkConsoleMessages } from '../framework/shared';
 
 import pushPoll from '../../../playground/mocks/data/idp/idx/authenticator-verification-okta-verify-push';
 import success from '../../../playground/mocks/data/idp/idx/success';
-import pushReject from '../../../playground/mocks/data/idp/idx/authenticator-verification-okta-verify-reject-push';
 
 const logger = RequestLogger(/challenge|challenge\/poll/,
   {
@@ -19,12 +18,6 @@ const pushSuccessMock = RequestMock()
   .respond(pushPoll)
   .onRequestTo('http://localhost:3000/idp/idx/challenge/poll')
   .respond(success);
-
-const pushRejectMock = RequestMock()
-  .onRequestTo('http://localhost:3000/idp/idx/introspect')
-  .respond(pushPoll)
-  .onRequestTo('http://localhost:3000/idp/idx/challenge/poll')
-  .respond(pushReject);
 
 const pushWaitMock = RequestMock()
   .onRequestTo('http://localhost:3000/idp/idx/introspect')
@@ -60,19 +53,6 @@ test
   });
 
 test
-  .requestHooks(pushRejectMock)('challenge okta verify with rejected push', async t => {
-    const challengeOktaVerifyPushPageObject = await setup(t);
-    await challengeOktaVerifyPushPageObject.waitForErrorBox();
-    const pageTitle = challengeOktaVerifyPushPageObject.getPageTitle();
-    await t.expect(pageTitle).contains('Get a push notification');
-    const errorBox = challengeOktaVerifyPushPageObject.getErrorBox();
-    await t.expect(errorBox.innerText).contains('You have chosen to reject this login.');
-    const resendPushBtn = challengeOktaVerifyPushPageObject.getResendPushButton();
-    await t.expect(resendPushBtn.value).contains('Resend push notification');
-    await t.expect(resendPushBtn.hasClass('link-button-disabled')).notOk();
-  });
-
-test
   .requestHooks(logger, pushSuccessMock)('challenge okta verify push request', async t => {
     await setup(t);
     const successPage = new SuccessPageObject(t);
@@ -93,31 +73,6 @@ test
     });
     await t.expect(answerRequestMethod).eql('post');
     await t.expect(answerRequestUrl).eql('http://localhost:3000/idp/idx/challenge/poll');
-  });
-
-test
-  .requestHooks(logger, pushRejectMock)('challenge okta verify resend push request', async t => {
-    const challengeOktaVerifyPushPageObject = await setup(t);
-    await challengeOktaVerifyPushPageObject.waitForErrorBox();
-    await challengeOktaVerifyPushPageObject.clickResendPushButton();
-
-    await t.expect(logger.count(() => true)).eql(2);
-    const { request: {
-      body: answerRequestBodyString,
-      method: answerRequestMethod,
-      url: answerRequestUrl,
-    }
-    } = logger.requests[1];
-    const answerRequestBody = JSON.parse(answerRequestBodyString);
-    await t.expect(answerRequestBody).eql({
-      authenticator: {
-        id: 'auteq0lLiL9o1cYoN0g4',
-        methodType: 'push',
-      },
-      stateHandle: '022P5Fd8jBy3b77XEdFCqnjz__5wQxksRfrAS4z6wP'
-    });
-    await t.expect(answerRequestMethod).eql('post');
-    await t.expect(answerRequestUrl).eql('http://localhost:3000/idp/idx/challenge');
   });
 
 test
